Fix jCenter build
