+++ apps/sasquatch/src/jcenterDependency/java/com/microsoft/appcenter/sasquatch/activities/BYOIUtils.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+package com.microsoft.appcenter.sasquatch.activities;
+
+/**
+ * TODO During release, delete this jCenter class version and move projectDependency to main folder.
+ */
+class BYOIUtils {
+
+    static void setAuthTokenListener() {
+    }
+
+    static void setAuthToken(@SuppressWarnings("unused") String authToken) {
+    }
+}
